implemented rbac on IDD coordinator suggestion endorsement

+++ pages/api/idd_coordinator_endorsement/[id]/index.ts
@@ -1,9 +1,6 @@
 import prisma from "@/prisma/client";
-import IDDCoordinatorEndorsementAbility from "@/services/ability/iDDCoordinatorEndorsementAbility";
 import getServerUser from "@/services/getServerUser";
 import logger from "@/services/logger";
-import { ForbiddenError } from "@casl/ability";
-import { accessibleBy } from "@casl/prisma";
 import { User } from "@prisma/client";
 import type { NextApiRequest, NextApiResponse } from "next";
 import * as Yup from "yup";
@@ -20,7 +17,6 @@ export default async function handler(
     logger.error(error);
     return res.status(401).json({ error: { message: "Unauthorized" } });
   }
-  const ability = IDDCoordinatorEndorsementAbility({ user });
 
   const getHandler = async () => {
     try {
@@ -35,7 +31,6 @@ export default async function handler(
         await prisma.iDDCoordinatorEndorsement.findFirstOrThrow({
           where: {
             AND: [
-              accessibleBy(ability).IDDCoordinatorEndorsement,
               {
                 id: {
                   equals: id,
@@ -62,12 +57,15 @@ export default async function handler(
 
       await validator.validate(req.query);
 
-      ForbiddenError.from(ability).throwUnlessCan(
-        "delete",
-        "IDDCoordinatorEndorsement"
-      );
-
       const { id } = validator.cast(req.query);
+
+      if (!user.isAdmin) {
+        return res.status(403).json({
+          error: {
+            message: "You are not allowed to perform this action",
+          },
+        });
+      }
 
       const IDDCoordinatorEndorsement =
         await prisma.iDDCoordinatorEndorsement.delete({
